Add attribute processing routine.
Attributes are now stored in sysdata.rda and processed as part of the NetCDF build process. To edit the attributes, modify the master table that is stored in data-raw/variable_attributes.csv, then variable_attributes.R in the data-raw directory. This will overwrite the variable_attributes object in sysdata.rda.